Adjustments to install scripts for both BBB and VM. - No longer directly replacing source.list, but adding architecture to existing repositories. - Added VU apt-get repository usage

+++ riaps-x86runtime/bootstrap.sh
@@ -57,14 +57,40 @@ user_func () {
 # Configure for cross functional compilation
 cross_setup() {
 	# Add armhf repositories
-	sudo apt-get install software-properties-common -y	
-	if grep -q 'http://ports.ubuntu.com/ubuntu-ports/' /etc/apt/sources.list ; 
-	then
-        echo "Armhf repositories are already included."
-    else
-        sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse"
-        sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial-updates main universe multiverse"
-    fi
+	sudo apt-get install software-properties-common apt-transport-https -y	
+    sudo add-apt-repository -r "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse" || true
+    sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse"
+    sudo add-apt-repository -r "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse" || true
+    sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse"    
+    
+    # Qualify the architectures for existing repositories trying to find armhf (which is not there) - this is due to issue installing later
+    # Need to figure out how not to need this (MM)
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial main restricted" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial main restricted"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates main restricted" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates main restricted"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial universe" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial universe"   
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates universe" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates universe"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial multiverse" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial multiverse"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates multiverse" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-updates multiverse"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-backports main restricted universe multiverse" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://us.archive.ubuntu.com/ubuntu/ xenial-backports main restricted universe multiverse" || true
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security main restricted" || true    
+    sudo add-apt-repository "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security main restricted"    
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security universe" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security universe"
+    sudo add-apt-repository -r "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security multiverse" || true
+    sudo add-apt-repository "deb [arch=amd64,i386] http://security.ubuntu.com/ubuntu xenial-security multiverse"
+    sudo add-apt-repository -r "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse" || true
+    sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial main universe multiverse"
+    sudo add-apt-repository -r "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial-updates main universe multiverse" || true
+    sudo add-apt-repository "deb [arch=armhf] http://ports.ubuntu.com/ubuntu-ports/ xenial-updates main universe multiverse"
+
+ 
     sudo dpkg --add-architecture armhf
     sudo apt-get update
     sudo apt-get install crossbuild-essential-armhf gdb-multiarch -y
@@ -123,12 +149,50 @@ curl_func () {
     echo "installed curl"
 }
 
-eclipse_func() {
-    sudo wget http://ftp.osuosl.org/pub/eclipse/technology/epp/downloads/release/neon/2/eclipse-java-neon-2-linux-gtk-x86_64.tar.gz
-    sudo -H -u $1 tar xfz eclipse-java-neon-2-linux-gtk-x86_64.tar.gz -C //home/$1/
 
-    sudo rm eclipse-java-neon-2-linux-gtk-x86_64.tar.gz
-    echo "installed eclipse"
+eclipse_shortcut() {
+   shortcut=/home/$1/Desktop/Eclipse.desktop
+   sudo -H -u $1 mkdir -p /home/$1/Desktop
+   sudo -H -u $1 cat <<EOT >$shortcut
+[Desktop Entry]
+Encoding=UTF-8
+Type=Application
+Name=Eclipse
+Name[en_US]=Eclipse
+Icon=/opt/eclipse/icon.xpm
+Exec=/opt/eclipse/eclipse -data /home/$1/workspace
+EOT
+
+   sudo chmod +x /home/$1/Desktop/Eclipse.desktop
+}
+
+eclipse_func() {
+	if [ ! -f "/home/$1/eclipse/eclipse" ]
+	then
+		if [ ! -f "/opt/eclipse/eclipse" ]
+    	then
+        	echo "eclipse not found"
+            sudo wget http://ftp.osuosl.org/pub/eclipse/technology/epp/downloads/release/neon/2/eclipse-java-neon-2-linux-gtk-x86_64.tar.gz
+            sudo tar xfz eclipse-java-neon-2-linux-gtk-x86_64.tar.gz -C /opt
+	    	#create eclipse shortcut
+			eclipse_shortcut $1
+			#install plugins
+            sudo /opt/eclipse/eclipse -clean  -consolelog  -noSplash -application org.eclipse.equinox.p2.director -repository http://pydev.org/updates -installIU "org.python.pydev.feature.feature.group, org.python.pydev.mylyn.feature.feature.group, org.python.pydev.feature.source.feature.group"    
+            #GIT
+            sudo /opt/eclipse/eclipse -clean  -consolelog  -noSplash -application org.eclipse.equinox.p2.director -repository http://download.eclipse.org/releases/neon/ -installIU "org.eclipse.egit.feature.group, org.eclipse.jgit.feature.group"	
+            #JSON Editor
+            sudo /opt/eclipse/eclipse -clean  -consolelog  -noSplash -application org.eclipse.equinox.p2.director -repository http://boothen.github.io/Json-Eclipse-Plugin/ -installIU "jsonedit-feature.feature.group"	
+            #Subclipse
+            sudo /opt/eclipse/eclipse -clean  -consolelog  -noSplash -application org.eclipse.equinox.p2.director -repository https://dl.bintray.com/subclipse/releases/subclipse/latest/ -installIU "org.tigris.subversion.subclipse.feature.group, net.java.dev.jna.feature.group, org.tigris.subversion.subclipse.mylyn.feature.feature.group, org.tigris.subversion.subclipse.graph.feature.feature.group, org.tigris.subversion.clientadapter.svnkit.feature.feature.group, org.tmatesoft.svnkit.feature.group"
+            sudo rm eclipse-java-neon-2-linux-gtk-x86_64.tar.gz
+            echo "installed eclipse"
+        else
+            echo "eclipse already installed at /opt/eclipse"
+        fi
+    else
+    echo "eclipse already installed at /home/riaps/eclipse"
+        
+    fi
 }
 
 install_redis () {
@@ -151,9 +215,8 @@ install_riaps() {
     # Add RIAPS repository
     sudo add-apt-repository -r "deb [arch=amd64] https://riaps.isis.vanderbilt.edu/aptrepo/ xenial main" || true
     sudo add-apt-repository "deb [arch=amd64] https://riaps.isis.vanderbilt.edu/aptrepo/ xenial main"
-
+    wget -qO - https://riaps.isis.vanderbilt.edu/keys/riapspublic.key | sudo apt-key add -
     sudo apt-get update
-    wget -qO - http://riaps.isis.vanderbilt.edu/keys/riapspublic.key | sudo apt-key add -
     ./riaps_install_amd64.sh
 }
